Move time intervals from GTIs to gti.py from utils.py
